Making it work with Live feeds
data-setup now can take a hash or string, but since it is interpolated
into a parameter on a video tag in single quotes, the inside must not
conflict with those single quotes, so hash is safer.

Updated Readme to have accurate examples (it used to have a misplaced
“}” so it was a bad example since it wouldn’t work.

Instead of assuming that the type of the source will always be
“video/…,” I changed it so that a string, not symbol is passed in.
Examples of other types:  rtmp/mp4, rtsp/mp4, application/x-mpegURL.

Since feeds that require flash need to actually have the video-js.swf
in the vendor/assets/flash folder, I created it and added the .swf file
there.  Also, put that requirement in the Readme that the variable
assignment for video-js.swf, which is critical for videojs to see the
swf.